♻️ refactor: Code Review 이름 변경
